bb-flasher-sd: flashing: Split tests
- Move tests to separate module. Was getting noisy.
